came to be able to get the score of own team and the remaining time of game.

+++ src/jp/ac/fit/asura/nao/communication/TeamInfo.java
@@ -55,7 +55,7 @@ public class TeamInfo {
 		return goalColor;
 	}
 
-	public short getScore() {
+	public byte getScore() {
 		return score;
 	}
